Allow autoOutdent to see new cursor pos on newline
Without this change, it's impossible for autoOutdent to outdent a new
line after it's been introduced. For example (the pipe character
represents the cursor):

1) function foo() {
2)    // ... |}

If newline is pressed, the "}" now on its own line will normally be
indented. If the mode wants to outdent it, it won't be possible
without this commit because the row number for line 2 is passed.